pyproject.toml
[tool.poetry]
name = "hexlet_code"
version = "0.1.0"
description = "Hexlet python project level 2"
authors = ["Kirill Chertkov <chertkov.k@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^6.2.4"
coverage = "^5.5"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.run:main"
